+++ lib/ffaker/guid.rb
@@ -5,7 +5,13 @@ module FFaker
     extend ModuleUtils
     extend self
 
+    # Because this method uses arbitrary hexadecimal characters it is likely to
+    # generate invalid UUIDs--UUIDs must have a version (1-8) at bits 48-51,
+    # and bits 64-65 must be 0b10.
+    #
+    # @deprecated Often generates invalid UUIDs. Use {UUID} instead.
     def guid
+      warn '[guid] is deprecated. Use the UUID module instead.'
       FFaker.hexify('########-####-####-####-############')
     end
   end